Create Leonardo ADA circle
This file is executed differently from "Leonardo ADA loop", on the other, to draw the round D line, we used a loop function which repeatetly drew 2 forward and 1 right, until it did a half circle. On this file, we just used a half circle (a circle which moves for only 180degrees).